Feat: 중복된 ID를 새로 저장하는 경우에 처리하는 로직 추가

+++ src/components/UIDefinitionChat.vue
@@ -101,13 +101,24 @@ export default {
          * 'Save' 버튼을 누를 경우, 최종 결과를 Supabase에 저장하기 위해서
          */
         async saveFormDefinition({id, name, html}){
+            const isNewSave = (this.$route.params.id !== id)
+            if(isNewSave) {
+                const isFormAlreadyExist = (await this.getData(`form_def/${id}`, { key: "id" }) !== null)
+                if(isFormAlreadyExist) {
+                    if(!confirm(`'${id}'는 이미 존재하는 폼 디자인 ID 입니다! 그래도 저장하시겠습니까?`))
+                        return
+                }
+            }
+
+
             await this.putObject("form_def", {
                 id, name, html,
                 messages: []
             });
             
+            
             alert("저장 완료!")
-            if(this.$route.params.id != id) {
+            if(isNewSave) {
                 this.$router.push(`/ui-definitions/${id}`)
             }
         },
@@ -121,6 +132,10 @@ export default {
             if (this.$route.params.id && this.$route.params.id != 'chat') {
                 path = `${this.path}/${this.$route.params.id}`
                 this.storedFormDefData = await this.getData(path, { key: "id" })
+                if(!this.storedFormDefData) {
+                    alert(`'${this.$route.params.id}' ID 를 가지는 폼 디자인 정보가 없습니다! 새 폼 만들기 화면으로 이동됩니다.`)
+                    this.$router.push(`/ui-definitions/chat`)
+                }
 
                 this.applyNewSrcToMashup(
                     this.loadHTMLToKEditorContent(this.storedFormDefData.html)
